Factorized glut code into glut_support.py
